requirements.txt
altgraph==0.17.4
anyio==4.1.0
astor==0.8.1
attrdict==2.0.1
Babel==2.13.1
bce-python-sdk==0.8.97
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli @ file:///C:/ci_311/brotli-split_1676435766766/work
cachetools==5.3.2
certifi @ file:///C:/b/abs_91u83siphd/croot/certifi_1700501720658/work/certifi
cffi @ file:///C:/b/abs_924gv1kxzj/croot/cffi_1700254355075/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
colorama==0.4.6
comtypes==1.3.0
contourpy==1.2.0
cryptography @ file:///C:/b/abs_f4do8t8jfs/croot/cryptography_1694444424531/work
cssselect==1.2.0
cssutils==2.9.0
customtkinter==5.2.2
cycler==0.12.1
Cython==3.0.6
darkdetect==0.8.0
decorator==5.1.1
dill==0.3.7
et-xmlfile==1.1.0
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
fire==0.5.0
Flask==3.0.0
flask-babel==4.0.0
fonttools==4.45.1
fsspec==2023.10.0
future==0.18.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna @ file:///C:/ci_311/idna_1676424932545/work
imageio==2.33.0
imgaug==0.4.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2 @ file:///C:/ci_311/jinja2_1676424968965/work
kiwisolver==1.4.5
lapx==0.5.9
lazy_loader==0.3
lmdb==1.4.1
lxml==4.9.3
MarkupSafe @ file:///C:/ci_311/markupsafe_1676424152318/work
matplotlib==3.8.2
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
networkx @ file:///C:/b/abs_e6gi1go5op/croot/networkx_1690562046966/work
numpy @ file:///C:/b/abs_9fu2cs2527/croot/numpy_and_numpy_base_1695830496596/work/dist/numpy-1.26.0-cp311-cp311-win_amd64.whl#sha256=ebb3a554edce1131a5b1b0d4afa8f980929b1226bebc4eb515ac2e29b44186d2
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
paddle-bfloat==0.1.7
paddleocr==2.7.0.3
paddlepaddle==2.5.2
pandas==2.1.3
pdf2docx==0.5.6
pefile==2023.2.7
Pillow @ file:///C:/b/abs_20ztcm8lgk/croot/pillow_1696580089746/work
premailer==3.10.0
protobuf==3.20.2
psutil==5.9.6
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodome==3.19.0
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2024.4
PyMuPDF @ file:///C:/Users/immer/.conda/envs/CevaExeGPU/Tools/MuPDF/PyMuPDF-1.20.2-cp311-cp311-win_amd64.whl#sha256=8969f3a8462352d8ddb3b679e87e85f84ae4fb49d2ef5d927563ceaca9a2aeaa
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing==3.1.1
pyqt6-plugins==6.4.2.2.3
PyQt6-Qt6==6.4.3
PyQt6-sip==13.6.0
pyqt6-tools==6.4.2.3.3
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
pysony==0.1.11
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pywinauto==0.6.8
PyYAML @ file:///C:/b/abs_782o3mbw7z/croot/pyyaml_1698096085010/work
qt6-applications==6.4.3.2.3
rapidfuzz==3.5.2
rarfile==4.1
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
scikit-image==0.22.0
scipy==1.11.4
seaborn==0.13.0
shapely==2.0.2-
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sympy @ file:///C:/ci_311_rebuilds/sympy_1679009400182/work
termcolor==2.3.0
thop==0.1.1.post2209072238
tifffile==2023.9.26
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
ttkthemes==3.2.2
typing_extensions @ file:///C:/b/abs_213vg2cd1l/croot/typing_extensions_1690297804941/work
tzdata==2023.3
ultralytics==8.0.220
urllib3 @ file:///C:/b/abs_9cmlsrm3ys/croot/urllib3_1698257595508/work
visualdl==2.5.3
Werkzeug==3.0.1
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
